Retitle IndexedDB “Basic concepts” article
The change retitles the IndexedDB “Basic concepts” article to
“IndexedDB key characteristics and basic terminology” — in order to make
more clear what the actual purpose of the article is, and what
information the article actually provides.

This change also does some streamlining of the article content — to
better align, and more clearly align, the article content with the
actual purpose of the article.

The scope of this article intentionally doesn’t include presenting an
example and walking the reader through it — that’s what the “Using
IndexedDB” tutorial is for. But in order for the reader to be able to
really understand the tutorial and find it useful, the reader first
needs to understand something about the key characteristics of
IndexedDB, along with some core terminology.

So this change repositions the article to more clearly align it with
that purpose.

Fixes https://github.com/mdn/content/issues/5340